Separate server and frontend error logic
