+++ app/src/components/OutputsTable/EditScenarioVars.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 import { BsPencil, BsX } from "react-icons/bs";
 import { api } from "~/utils/api";
 import { useExperiment, useHandledAsyncCallback, useScenarioVars } from "~/utils/hooks";
-import { maybeReportError } from "~/utils/standardResponses";
+import { maybeReportError } from "~/utils/errorHandling/maybeReportError";
 import { FloatingLabelInput } from "./FloatingLabelInput";
 
 export const ScenarioVar = ({
