+++ app/src/utils/errorHandling/maybeReportError.ts
@@ -1,16 +1,5 @@
 import { toast } from "~/theme/ChakraThemeProvider";
-
-export function error(message: string): { status: "error"; message: string } {
-  return {
-    status: "error",
-    message,
-  };
-}
-export function success<T>(payload: T): { status: "success"; payload: T };
-export function success(payload?: undefined): { status: "success"; payload: undefined };
-export function success<T>(payload?: T) {
-  return { status: "success", payload };
-}
+import { type error, type success } from "./standardResponses";
 
 type SuccessType<T> = ReturnType<typeof success<T>>;
 type ErrorType = ReturnType<typeof error>;
